requirements.txt
# ============================================
# DDFD-FaceRec 核心依赖
# ============================================

# --------------------------------------------
# 深度学习框架
# --------------------------------------------
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# --------------------------------------------
# 图像处理
# --------------------------------------------
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
scikit-image>=0.21.0

# --------------------------------------------
# 科学计算
# --------------------------------------------
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# --------------------------------------------
# 特征检索
# --------------------------------------------
hnswlib>=0.7.0
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # GPU 版本，按需安装

# --------------------------------------------
# 模型导出与推理
# --------------------------------------------
onnx>=1.13.0
onnxruntime-gpu>=1.15.0
onnxsim>=0.4.3
# tensorrt>=8.6.0  # 需从 NVIDIA 官方安装
# openvino>=2023.0.0  # 按需安装

# --------------------------------------------
# API 服务
# --------------------------------------------
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.1.0

# --------------------------------------------
# 数据库与缓存
# --------------------------------------------
redis>=4.6.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# --------------------------------------------
# 工具库
# --------------------------------------------
tqdm>=4.65.0
yaml>=6.0
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0
loguru>=0.7.0
tensorboard>=2.13.0
wandb>=0.15.0

# --------------------------------------------
# 测试
# --------------------------------------------
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # API 测试

# --------------------------------------------
# 代码质量
# --------------------------------------------
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0

# --------------------------------------------
# 容器化
# --------------------------------------------
# docker>=6.0.0  # 按需安装

# --------------------------------------------
# 其他
# --------------------------------------------
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
fire>=0.5.0  # 命令行工具
